FREEPBX-23743
Add storage support so you can turn it on and off without having to remove it.
Block access to storage if disabled.
Fix exception error that occurs if you specify a driver that does not exist.
Used show_help to generate msg What is File Store?
Replace the call "$this->FreePBX->Filestore->" for "$this->".
Extract the html code inside getRightNav to an external file in the views folder.
Now after modifying a storage we return to the list of storages and the tab corresponding to the type of driver in which we were is activated.